+++ core/src/main/java/com/lukakordzaia/core/domain/repository/databaserepository/DatabaseRepository.kt
@@ -5,7 +5,7 @@ import com.lukakordzaia.core.database.continuewatchingdb.ContinueWatchingRoom
 
 interface DatabaseRepository {
     fun getContinueWatchingFromRoom(): LiveData<List<ContinueWatchingRoom>>
-    suspend fun deleteSingleContinueWatchingFromRoom(titleId: Int): Long
+    suspend fun deleteSingleContinueWatchingFromRoom(titleId: Int): Int
     fun getSingleContinueWatchingFromRoom(titleId: Int): LiveData<ContinueWatchingRoom>
     suspend fun insertContinueWatchingInRoom(continueWatchingRoom: ContinueWatchingRoom)
     suspend fun deleteAllFromRoom()
