+++ core/src/main/java/com/lukakordzaia/core/domain/repository/databaserepository/DefaultDatabaseRepository.kt
@@ -9,7 +9,7 @@ class DefaultDatabaseRepository(private val database: StreamFlowDatabase): Datab
         return database.continueWatchingDao().getContinueWatchingFromRoom()
     }
 
-    override suspend fun deleteSingleContinueWatchingFromRoom(titleId: Int): Long {
+    override suspend fun deleteSingleContinueWatchingFromRoom(titleId: Int): Int {
         return database.continueWatchingDao().deleteSingleContinueWatchingFromRoom(titleId)
     }
